chapter-07/requirements.txt
openai==1.98.0
python-dotenv==1.1.1
scipy==1.16.1
